requirements.txt
Django==2.2.7
django_mptt==0.13.2

django-crispy-forms==1.12.0
django-js-asset==1.2.2

npm==0.1.1
optional-django==0.1.0
Pillow==8.3.1
pytz==2019.3
sqlparse==0.3.0



